backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-cloud-vision>=3.7.2
Pillow>=10.4.0
python-jose>=3.3.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
